Refactor to Routed Compnents to a Page Directory
