+++ src/App.js
@@ -3,17 +3,17 @@ import React from 'react'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
 
-import Analytics from '../components/Analytics'
-import BetaBanner from '../components/BetaBanner'
-import ClearCacheBtn from '../components/ClearCacheBtn'
-import Disclaimer from '../components/Disclaimer'
-import Feedback from '../components/Feedback'
-import Footer from '../components/Footer'
-import Glossary from '../components/Glossary'
-import Header from '../components/Header'
-import SkipContent from '../components/SkipContent'
-import * as feedbackActions from '../actions/feedback'
-import * as glossaryActions from '../actions/glossary'
+import Analytics from './components/Analytics'
+import BetaBanner from './components/BetaBanner'
+import ClearCacheBtn from './components/ClearCacheBtn'
+import Disclaimer from './components/Disclaimer'
+import Feedback from './components/Feedback'
+import Footer from './components/Footer'
+import Glossary from './components/Glossary'
+import Header from './components/Header'
+import SkipContent from './components/SkipContent'
+import * as feedbackActions from './actions/feedback'
+import * as glossaryActions from './actions/glossary'
 
 const isProd = process.env.NODE_ENV === 'production'
