+++ src/pages/About.js
@@ -4,11 +4,11 @@ import Helmet from 'react-helmet'
 import { Link } from 'react-router'
 import reduceEntries from 'reduce-entries'
 
-import SharingTags from './SharingTags'
-import Term from './Term'
-import UsaMap from './UsaMap'
+import SharingTags from '../components/SharingTags'
+import Term from '../components/Term'
+import UsaMap from '../components/UsaMap'
 
-import DownloadDataBtn from './DownloadDataBtn'
+import DownloadDataBtn from '../components/DownloadDataBtn'
 import { showFeedback } from '../actions/feedback'
 import participation from '../util/participation'
 import { data as usaData, nationalKey } from '../util/usa'
