+++ src/pages/Explorer.js
@@ -3,12 +3,12 @@ import PropTypes from 'prop-types'
 import React from 'react'
 import Helmet from 'react-helmet'
 
-import AboutTheData from './AboutTheData'
+import AboutTheData from '../components/AboutTheData'
 import AgencyChartContainer from '../containers/AgencyChartContainer'
 import ExplorerHeaderContainer from '../containers/ExplorerHeaderContainer'
 import NibrsContainer from '../containers/NibrsContainer'
 import NotFound from './NotFound'
-import SharingTags from './SharingTags'
+import SharingTags from '../components/SharingTags'
 import SidebarContainer from '../containers/SidebarContainer'
 import SparklineContainer from '../containers/SparklineContainer'
 import TrendContainer from '../containers/TrendContainer'
